gift card provider now extends invoice.ts, moved email input above details card, new invoice folder inside integration pages
