+++ src/pages/integrations/invoice/confirm-invoice/confirm-invoice.html
@@ -86,6 +86,32 @@
         </div>
       </button>
 
+      <div>
+        <ion-item-divider> {{ 'Refund Email' | translate }} </ion-item-divider>
+        <ion-item>
+          <ion-input
+            placeholder="{{'Enter an email address' | translate}}"
+            [(ngModel)]="email"
+            name="email"
+            type="email"
+            [disabled]="
+              !!(this.buyerProvidedEmail || this.merchantProvidedEmail)
+            "
+          ></ion-input>
+        </ion-item>
+      </div>
+
+      <label-tip *ngIf="!isValidEmail()" type="danger">
+        <span label-tip-title translate
+          ><strong>A refund email is required.</strong></span
+        >
+        <div label-tip-body translate>
+          If there is an issue with this payment, or a refund needs to be made,
+          we will contact you at this address.<br />
+          <a (click)="openPrivacyPolicy()" translate>View Privacy Policy</a>
+        </div>
+      </label-tip>
+
       <div *ngIf="totalAmountStr">
         <ion-item-divider> {{ 'Details' | translate }} </ion-item-divider>
         <ion-item>
@@ -134,32 +160,6 @@
           </ion-note>
         </ion-item>
       </div>
-
-      <div>
-        <ion-item-divider> {{ 'Refund Email' | translate }} </ion-item-divider>
-        <ion-item>
-          <ion-input
-            placeholder="{{'Enter an email address' | translate}}"
-            [(ngModel)]="email"
-            name="email"
-            type="email"
-            [disabled]="
-              !!(this.buyerProvidedEmail || this.merchantProvidedEmail)
-            "
-          ></ion-input>
-        </ion-item>
-      </div>
-
-      <label-tip *ngIf="!isValidEmail()" type="danger">
-        <span label-tip-title translate
-          ><strong>A refund email is required.</strong></span
-        >
-        <div label-tip-body translate>
-          If there is an issue with this payment, or a refund needs to be made,
-          we will contact you at this address.<br />
-          <a (click)="openPrivacyPolicy()" translate>View Privacy Policy</a>
-        </div>
-      </label-tip>
       <ion-item-divider></ion-item-divider>
     </ion-item-group>
   </ion-list>
