+++ src/providers/gift-card/gift-card.ts
@@ -1,4 +1,4 @@
-import { HttpClient, HttpHeaders } from '@angular/common/http';
+import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { ImageLoader } from 'ionic-image-loader';
 import * as _ from 'lodash';
@@ -11,12 +11,9 @@ import { promiseSerial } from '../../utils';
 import { ConfigProvider } from '../config/config';
 import { EmailNotificationsProvider } from '../email-notifications/email-notifications';
 import { HomeIntegrationsProvider } from '../home-integrations/home-integrations';
+import { InvoiceProvider } from '../invoice/invoice';
 import { Logger } from '../logger/logger';
-import {
-  GiftCardMap,
-  Network,
-  PersistenceProvider
-} from '../persistence/persistence';
+import { GiftCardMap, PersistenceProvider } from '../persistence/persistence';
 import { TimeProvider } from '../time/time';
 import {
   ApiCardConfig,
@@ -28,15 +25,7 @@ import {
 } from './gift-card.types';
 
 @Injectable()
-export class GiftCardProvider {
-  credentials: {
-    NETWORK: Network;
-    BITPAY_API_URL: string;
-  } = {
-    NETWORK: Network.livenet,
-    BITPAY_API_URL: 'https://bitpay.com'
-  };
-
+export class GiftCardProvider extends InvoiceProvider {
   availableCardsPromise: Promise<CardConfig[]>;
   availableCardMapPromise: Promise<{ [name: string]: CardConfig }>;
 
@@ -58,18 +47,9 @@ export class GiftCardProvider {
     public persistenceProvider: PersistenceProvider,
     public timeProvider: TimeProvider
   ) {
+    super(emailNotificationsProvider, http, logger, persistenceProvider);
     this.logger.debug('GiftCardProvider initialized');
     this.setCredentials();
-  }
-
-  getNetwork() {
-    return this.credentials.NETWORK;
-  }
-
-  setCredentials() {
-    if (this.getNetwork() === Network.testnet) {
-      this.credentials.BITPAY_API_URL = 'https://test.bitpay.com';
-    }
   }
 
   async getCardConfig(cardName: string) {
@@ -350,42 +330,6 @@ export class GiftCardProvider {
     );
   }
 
-  async createBitpayInvoice(data) {
-    const dataSrc = {
-      brand: data.cardName,
-      currency: data.currency,
-      amount: data.amount,
-      clientId: data.uuid,
-      email: data.email,
-      transactionCurrency: data.buyerSelectedTransactionCurrency
-    };
-    const url = `${this.getApiPath()}/pay`;
-    const headers = new HttpHeaders({
-      'Content-Type': 'application/json'
-    });
-    const cardOrder = await this.http
-      .post(url, dataSrc, { headers })
-      .toPromise()
-      .catch(err => {
-        this.logger.error('BitPay Create Invoice: ERROR', JSON.stringify(data));
-        throw err;
-      });
-    this.logger.info('BitPay Create Invoice: SUCCESS');
-    return cardOrder as { accessKey: string; invoiceId: string };
-  }
-
-  public async getBitPayInvoice(id: string) {
-    const res: any = await this.http
-      .get(`${this.credentials.BITPAY_API_URL}/invoices/${id}`)
-      .toPromise()
-      .catch(err => {
-        this.logger.error('BitPay Get Invoice: ERROR ' + err.error.message);
-        throw err.error.message;
-      });
-    this.logger.info('BitPay Get Invoice: SUCCESS');
-    return res.data;
-  }
-
   private checkIfCardNeedsUpdate(card: GiftCard) {
     if (!card.invoiceId) {
       return false;
@@ -531,35 +475,6 @@ export class GiftCardProvider {
     return this.availableCardsPromise;
   }
 
-  getApiPath() {
-    return `${this.credentials.BITPAY_API_URL}/gift-cards`;
-  }
-
-  public emailIsValid(email: string): boolean {
-    const validEmail = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/.test(
-      email
-    );
-    return validEmail;
-  }
-
-  public storeEmail(email: string): void {
-    this.setUserInfo({ email });
-  }
-
-  public getUserEmail(): Promise<string> {
-    return this.persistenceProvider
-      .getGiftCardUserInfo()
-      .then(data => {
-        if (_.isString(data)) {
-          data = JSON.parse(data);
-        }
-        return data && data.email
-          ? data.email
-          : this.emailNotificationsProvider.getEmailIfEnabled();
-      })
-      .catch(_ => {});
-  }
-
   public async preloadImages(): Promise<void> {
     const supportedCards = await this.getSupportedCards();
     const imagesPerCard = supportedCards
@@ -569,10 +484,6 @@ export class GiftCardProvider {
       Promise.all(images.map(i => this.imageLoader.preload(i)))
     );
     await promiseSerial(fetchBatches);
-  }
-
-  private setUserInfo(data: any): void {
-    this.persistenceProvider.setGiftCardUserInfo(JSON.stringify(data));
   }
 
   public register() {
